requirements.txt
# Indonesia EdTech Lead Gen Engine
# Core dependencies

# Environment & Config
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# HTTP Clients
httpx>=0.25.0
aiohttp>=3.9.0

# Web Scraping
crawl4ai>=0.3.0
playwright>=1.40.0

# PDF Extraction
pymupdf>=1.23.0

# LLM & Extraction (OpenRouter compatible)
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.0.0

# Data Export
pandas>=2.1.0
openpyxl>=3.1.0

# Async support
aiofiles>=23.2.0

# Progress bar
tqdm>=4.66.0

# Rich console output
rich>=13.7.0

# Email validation
email-validator>=2.1.0
dnspython>=2.4.0

# FastAPI backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Supabase
supabase>=2.0.0
postgrest>=0.13.0

